requirements.txt
Django==2.2.7
djangorestframework==3.10.3
parameterized==0.7.1